+++ tests/unit/test_utils.py
@@ -204,6 +204,39 @@ def test_format_list_collect():
     assert 4 == len(all_fields)
 
 
+def test_formatting():
+    from ldap2pg.utils import FormatList, make_format_vars
+
+    entry = (
+        'dn',
+        {
+            'dn': ['dn'],
+            'member': [
+                'cn=m0',
+                'cn=m1',
+            ],
+        },
+        {'member': [
+            ('cn=m0', {
+                'dn': ['cn=m0'],
+                'mail': ['m0@toto'],
+            }, {}),
+            ('cn=m1', {
+                'dn': ['cn=m1'],
+                'mail': ['m1@toto'],
+            }, {}),
+        ]},
+    )
+    formats = FormatList.factory(['{member.cn} <{member.mail}>'])
+    values = {
+        'member.cn': ['m0', 'm1'],
+        'member.mail': ['m0@toto', 'm1@toto']
+    }
+    vars_ = make_format_vars(formats.fields, entry[0], values)
+    strings = list(formats.expand(vars_))
+    assert ['m0 <m0@toto>', 'm1 <m1@toto>'] == strings
+
+
 def test_user_error_wrap():
     from ldap2pg.utils import UserError