+++ docs/changelog.md
@@ -11,6 +11,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 
 - Fails if configuration file is not found.
 - Refuse empty configuration file.
+- Refuse undefined `sync_map`.
 
 
 # ldap2pg 5.2