+++ docs/changelog.md
@@ -7,6 +7,13 @@ follow [merged Pull request
 pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Amerged).
 
 
+# Unreleased
+
+- Fails if configuration file is not found.
+- Refuse empty configuration file.
+- Refuse undefined `sync_map`.
+
+
 # ldap2pg 5.2
 
 **Attention!** This release has some behaviour changes. Some silented errors