+++ docs/changelog.md
@@ -12,6 +12,8 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 - Fails if configuration file is not found.
 - Refuse empty configuration file.
 - Refuse undefined `sync_map`.
+- Fix join when multiple entries are returned.
+- Fix using multiple attributes from joined entries.
 
 
 # ldap2pg 5.2