+++ docs/changelog.md
@@ -9,12 +9,13 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 
 # Unreleased
 
-- Fails if configuration file is not found.
-- Refuse empty configuration file.
-- Refuse undefined `sync_map`.
 - Fix join when multiple entries are returned.
 - Fix using multiple attributes from joined entries.
 - Fix comment error with generated comment.
+- Fails if configuration file is not found.
+- Refuse empty configuration file.
+- Refuse undefined `sync_map`.
+- Update sample configuration for readability and general use.
 
 
 # ldap2pg 5.2