+++ docs/changelog.md
@@ -7,7 +7,7 @@ follow [merged Pull request
 pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Amerged).
 
 
-# Unreleased
+# ldap2pg 5.3
 
 - Fix join when multiple entries are returned.
 - Fix using multiple attributes from joined entries.